+++ 05-mixnets/core/abstractions.py
@@ -1,3 +1,4 @@
+from models import Message
 class BaseIO:
     def __init__(self):
         self.on_message = None
@@ -20,3 +21,40 @@ class BaseAddressResolver:
 class BaseMessageSender:
     def send_message(self, type: str, payload: [bytes], receiver: str):
         raise
+
+
+class BaseRouter:
+    def __init__(self):
+        pass
+
+    def _schedule_next_announce(self):
+        pass
+
+    def announce(self):
+        pass
+
+    def resend_announce(self, sender: str, message: Message):
+        pass
+
+    def receive_message(self, msg: [bytes], sender: str):
+        pass
+
+    def send_message(self, msg: bytes, sender_public_key: str):
+        pass
+
+    def find_announce_match(self, target_hash, address):
+        pass
+
+class BaseMessageInput:
+    def __init__(self, router: BaseRouter):
+        self.router = router
+
+    def send_message(self, msg: bytes, public_key: str):
+        self.router.send_message(msg, public_key)
+
+class BaseMessageOutput:
+    def __init__(self):
+        pass
+
+    def accept_message(self, bytes):
+        pass
